Add user ranking badges to title on user profile page
Show in the title of the user profile page SPANs with
CSS class names to represent the user's rankings, so
they can be represented in the UI by "badges".

The base user serializer can now generates a property
`user_badge_css_classes` that includes a list of CSS
class names generated by a function at the path
configured by the optional setting
`MISAGO_USER_BADGES_FUNCTION`. If this setting is
defined, the function is called and its results
returned directly.