Avoid broken signature render in post preview
The preview widget was failing due to trying to
show user signatures in the preview modal pop-up
when this data was not available for new posts
(as opposed to editing existing posts).

This change checks whether existing post data
is available before using it to render signature
content.

See https://github.com/ixc/iwc-forum/issues/57